Enable stateless XNNPACK linear.
The optimal solution to use XNNPACK is to separate operator creation
from execution - also called prepacking the weights.  If we have done
our job properly, JIT must have caught and replaced nn.Linear on mobile
with the prepacked versions.  Still, if we somehow end up in
at::native::linear for whatever reason, it is still more efficient to go
through XNNPACK than the alternatives of at::addmm or at::matmul.

ghstack-source-id: a59d316d9b0b402448cb0c3117b35f1e9c073a65
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35791